Move remove_from_selfserve_* functions from installer to manifestutils to prevent an import loop in removepackages
