+++ code/client/munkilib/updatecheck/manifestutils.py
@@ -266,6 +266,50 @@ def get_manifest_value_for_key(manifestpath, keyname):
         return None
 
 
+def remove_from_selfserve_section(itemname, section):
+    """Remove the given itemname from the self-serve manifest's
+    managed_uninstalls list"""
+    display.display_debug1(
+        "Removing %s from SelfSeveManifest's %s...", itemname, section)
+    selfservemanifest = os.path.join(
+        prefs.pref('ManagedInstallDir'), 'manifests', 'SelfServeManifest')
+    if not os.path.exists(selfservemanifest):
+        # SelfServeManifest doesn't exist, bail
+        display.display_debug1("%s doesn't exist.", selfservemanifest)
+        return
+    try:
+        plist = FoundationPlist.readPlist(selfservemanifest)
+    except FoundationPlist.FoundationPlistException, err:
+        # SelfServeManifest is broken, bail
+        display.display_debug1(
+            "Error reading %s: %s", selfservemanifest, err)
+        return
+    # make sure the section is in the plist
+    if section in plist:
+        # filter out our item
+        plist[section] = [
+            item for item in plist[section] if item != itemname
+        ]
+        try:
+            FoundationPlist.writePlist(plist, selfservemanifest)
+        except FoundationPlist.FoundationPlistException, err:
+            display.display_debug1(
+                "Error writing %s: %s", selfservemanifest, err)
+
+
+def remove_from_selfserve_installs(itemname):
+    """Remove the given itemname from the self-serve manifest's
+    managed_installs list"""
+    remove_from_selfserve_section(itemname, 'managed_installs')
+
+
+def remove_from_selfserve_uninstalls(itemname):
+    """Remove the given itemname from the self-serve manifest's
+    managed_uninstalls list"""
+    # pylint: disable=invalid-name
+    remove_from_selfserve_section(itemname, 'managed_uninstalls')
+
+
 # module globals
 _MANIFESTS = {}
