+++ code/client/removepackages
@@ -66,7 +66,7 @@ def main():
 
     # check to see if we're root
     if os.geteuid() != 0:
-        print sys.stderr, "You must run this as root!"
+        print >> sys.stderr, "You must run this as root!"
         exit(-1)
 
     # set the display globals
